Adding ruby-endpoints to the non-clobber directories
This commit is to ensure the jekyll build doesn't wipe out
all the ruby endpoints we have in `~/public_html/`. See
ECAConcordia/IT#2 for a description of what was happening
and the fixes to the `.htaccess`es that were required.

So, with any luck, our commit hook won't break after this push.